dom/html/test/test_bug401160.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=401160
-->
<head>
  <title>Test for Bug 401160</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=401160">Mozilla Bug 401160</a>
<label id="label" contenteditable="true"><legend></legend><div></div></label>

<pre id="test">
<script type="text/javascript">

function do_test() {
  document.getElementById('label').focus();
  ok(true, "This is crash test - the test succeeded if we reach this line")
}

do_test();

</script>
</pre>
</body>
</html>
